[JF] add jf-control-app
Create a jf-control-app built around existing chip-tool. Extracted
only the required functionality from chip-tool and duplicated files
inside jf-control-app.

For now, the app plays the role of both of a commissioner and also
of an administrator.

Later, jf-control-add plays will play the role of a commissioner
only which interacts with the jf-admin-app (administrator role).

Signed-off-by: Doru Gucea <doru-cristian.gucea@nxp.com>