examples/jf-control-app: Fix CLI commands reporting incorrect number of arguments
The new arguments ("--commissioner-name <name>") added automatically
to the "Commands" class are affecting all CLI commands, not just the
cluster commands. For non-cluster commands however (such as
"rpc status", or "rpc connect"), these arguments are not meaningful
and not necessary either.

The solution here would be to relocate the automatic definition of
these arguments lower in the class hierarchy, to the "CHIPCommand"
class.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>